% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sitter.R
\name{dom_toml}
\alias{dom_toml}
\title{Print the document tree of a TOML file or string}
\usage{
dom_toml(
  file = NULL,
  text = NULL,
  ranges = NULL,
  fail_on_parse_error = TRUE,
  options = NULL
)
}
\arguments{
\item{file}{Path of a TOML file. Use either \code{file} or \code{text}.}

\item{text}{TOML string. Use either \code{file} or \code{text}.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}

\item{fail_on_parse_error}{Logical, whether to error if there are
parse errors in the toml document. Default is \code{TRUE}.}

\item{options}{List of options, see \code{\link[=tstoml_options]{tstoml_options()}}. This argument
must be named and cannot be abbreviated.}
}
\description{
Print the document tree of a TOML file or string
}
