% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-toml.R
\name{load_toml}
\alias{load_toml}
\title{Parse a TOML file or string into a tstoml object}
\usage{
load_toml(file = NULL, text = NULL, ranges = NULL, options = NULL)
}
\arguments{
\item{file}{Path of a TOML file. Use either \code{file} or \code{text}.}

\item{text}{TOML string. Use either \code{file} or \code{text}.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}

\item{options}{List of options, see \code{\link[=tstoml_options]{tstoml_options()}}. This argument
must be named and cannot be abbreviated.}
}
\value{
A tstoml object.
}
\description{
Parse a TOML file or string into a tstoml object
}
