% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sitter.R
\name{query_toml}
\alias{query_toml}
\title{Run tree-sitter queries on a TOML file or string}
\usage{
query_toml(file = NULL, text = NULL, query, ranges = NULL)
}
\arguments{
\item{file}{Path of a TOML file. Use either \code{file} or \code{text}.}

\item{text}{TOML string. Use either \code{file} or \code{text}.}

\item{query}{Character string, the tree-sitter query to run.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}
}
\value{
A list with entries \code{patterns} and \code{matched_captures}.
\code{patterns} contains information about all patterns in the queries and
it is a data frame with columns: \code{id}, \code{name}, \code{pattern}, \code{match_count}.
\code{matched_captures} contains information about all matches, and it has
columns \code{id}, \code{pattern}, \code{match}, \code{start_byte}, \code{end_byte}, \code{start_row},
\code{start_column}, \code{end_row}, \code{end_column}, \code{name}, \code{code}. The \code{pattern}
column of \code{matched_captured} refers to the \code{id} column of \code{patterns}.
}
\description{
See https://tree-sitter.github.io/tree-sitter/ on writing tree-sitter
queries.
}
