% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstoml-package.R
\name{quickstart}
\alias{quickstart}
\title{tstoml quickstart}
\description{
tstoml quickstart
}
\details{
\subsection{Create a tstoml object}{

Create a tstoml object from a string:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{txt <- r"(
# This is a TOML document

title = "TOML Example"

[owner]
name = "Tom Preston-Werner"
dob = 1979-05-27T07:32:00-08:00

[database]
enabled = true
ports = [ 8000, 8001, 8002 ]
data = [ ["delta", "phi"], [3.14] ]
temp_targets = \{ cpu = 79.5, case = 72.0 \}

[servers]

[servers.alpha]
ip = "10.0.0.1"
role = "frontend"

[servers.beta]
ip = "10.0.0.2"
role = "backend"
)"
toml <- ts_parse_toml(text = txt)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#>                                                                                 
</pre></div>
}}


Pretty print a tstoml object:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{toml
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (24 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span>                                                                           
#> <span style="color: #6F7783;"> 2 | </span># This is a TOML document                                                  
#> <span style="color: #6F7783;"> 3 | </span>                                                                           
#> <span style="color: #6F7783;"> 4 | </span>title = "TOML Example"                                                     
#> <span style="color: #6F7783;"> 5 | </span>                                                                           
#> <span style="color: #6F7783;"> 6 | </span>[owner]                                                                    
#> <span style="color: #6F7783;"> 7 | </span>name = "Tom Preston-Werner"                                                
#> <span style="color: #6F7783;"> 8 | </span>dob = 1979-05-27T07:32:00-08:00                                            
#> <span style="color: #6F7783;"> 9 | </span>                                                                           
#> <span style="color: #6F7783;">10 | </span>[database]                                                                 
#> <span style="color: #6F7783;">ℹ 14 more lines</span>                                                                 
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}

}

\subsection{Select elements in a tstoml object}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(toml, "owner")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (24 lines, 1 selected element)</span>                                           
#>   <span style="color: #6F7783;">...</span>                                                                           
#>   <span style="color: #6F7783;"> 3  | </span>                                                                        
#>   <span style="color: #6F7783;"> 4  | </span>title = "TOML Example"                                                  
#>   <span style="color: #6F7783;"> 5  | </span>                                                                        
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 6  | </span><span style="color: #66C2CD;">[owner]</span>                                                                 
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 7  | </span><span style="color: #66C2CD;">name = "Tom Preston-Werner"</span>                                             
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 8  | </span><span style="color: #66C2CD;">dob = 1979-05-27T07:32:00-08:00</span>                                         
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 9  | </span>                                                                        
#>   <span style="color: #6F7783;">10  | </span>[database]                                                              
#>   <span style="color: #6F7783;">11  | </span>enabled = true                                                          
#>   <span style="color: #6F7783;">12  | </span>ports = [ 8000, 8001, 8002 ]                                            
#>   <span style="color: #6F7783;">...</span>                                                                           
</pre></div>
}}


Select element(s) inside elements:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(toml, "owner", "name")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (24 lines, 1 selected element)</span>                                           
#>   <span style="color: #6F7783;">...</span>                                                                           
#>   <span style="color: #6F7783;"> 4  | </span>title = "TOML Example"                                                  
#>   <span style="color: #6F7783;"> 5  | </span>                                                                        
#>   <span style="color: #6F7783;"> 6  | </span>[owner]                                                                 
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 7  | </span>name = <span style="color: #66C2CD;">"Tom Preston-Werner"</span>                                             
#>   <span style="color: #6F7783;"> 8  | </span>dob = 1979-05-27T07:32:00-08:00                                         
#>   <span style="color: #6F7783;"> 9  | </span>                                                                        
#>   <span style="color: #6F7783;">10  | </span>[database]                                                              
#>   <span style="color: #6F7783;">...</span>                                                                           
</pre></div>
}}


Select element(s) of an array:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(toml, "database", "ports", 1:2)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (24 lines, 2 selected elements)</span>                                          
#>   <span style="color: #6F7783;">...</span>                                                                           
#>   <span style="color: #6F7783;"> 9  | </span>                                                                        
#>   <span style="color: #6F7783;">10  | </span>[database]                                                              
#>   <span style="color: #6F7783;">11  | </span>enabled = true                                                          
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;">12  | </span>ports = [ <span style="color: #66C2CD;">8000</span>, <span style="color: #66C2CD;">8001</span>, 8002 ]                                            
#>   <span style="color: #6F7783;">13  | </span>data = [ ["delta", "phi"], [3.14] ]                                     
#>   <span style="color: #6F7783;">14  | </span>temp_targets = \{ cpu = 79.5, case = 72.0 \}                              
#>   <span style="color: #6F7783;">15  | </span>                                                                        
#>   <span style="color: #6F7783;">...</span>                                                                           
</pre></div>
}}


Select multiple keys from a table:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(toml, "owner", c("name", "dob"))
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (24 lines, 2 selected elements)</span>                                          
#>   <span style="color: #6F7783;">...</span>                                                                           
#>   <span style="color: #6F7783;"> 4  | </span>title = "TOML Example"                                                  
#>   <span style="color: #6F7783;"> 5  | </span>                                                                        
#>   <span style="color: #6F7783;"> 6  | </span>[owner]                                                                 
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 7  | </span>name = <span style="color: #66C2CD;">"Tom Preston-Werner"</span>                                             
#> <span style="background-color: #66C2CD;">&gt;</span> <span style="color: #6F7783;"> 8  | </span>dob = <span style="color: #66C2CD;">1979-05-27T07:32:00-08:00</span>                                         
#>   <span style="color: #6F7783;"> 9  | </span>                                                                        
#>   <span style="color: #6F7783;">10  | </span>[database]                                                              
#>   <span style="color: #6F7783;">11  | </span>enabled = true                                                          
#>   <span style="color: #6F7783;">...</span>                                                                           
</pre></div>
}}

}

\subsection{Delete elements}{

Delete selected elements:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(toml, "owner", "name") |> ts_tree_delete()
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (22 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span># This is a TOML document                                                  
#> <span style="color: #6F7783;"> 2 | </span>                                                                           
#> <span style="color: #6F7783;"> 3 | </span>title = "TOML Example"                                                     
#> <span style="color: #6F7783;"> 4 | </span>                                                                           
#> <span style="color: #6F7783;"> 5 | </span>[owner]                                                                    
#> <span style="color: #6F7783;"> 6 | </span>dob = 1979-05-27T07:32:00-08:00                                            
#> <span style="color: #6F7783;"> 7 | </span>                                                                           
#> <span style="color: #6F7783;"> 8 | </span>[database]                                                                 
#> <span style="color: #6F7783;"> 9 | </span>enabled = true                                                             
#> <span style="color: #6F7783;">10 | </span>ports = [ 8000, 8001, 8002 ]                                               
#> <span style="color: #6F7783;">ℹ 12 more lines</span>                                                                 
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}

}

\subsection{Insert elements}{

Insert a key-value pair into the document:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_insert(toml, key = "new_key", "new_value")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (25 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span># This is a TOML document                                                  
#> <span style="color: #6F7783;"> 2 | </span>                                                                           
#> <span style="color: #6F7783;"> 3 | </span>title = "TOML Example"                                                     
#> <span style="color: #6F7783;"> 4 | </span>                                                                           
#> <span style="color: #6F7783;"> 5 | </span>                                                                           
#> <span style="color: #6F7783;"> 6 | </span>new_key = "new_value"                                                      
#> <span style="color: #6F7783;"> 7 | </span>[owner]                                                                    
#> <span style="color: #6F7783;"> 8 | </span>name = "Tom Preston-Werner"                                                
#> <span style="color: #6F7783;"> 9 | </span>dob = 1979-05-27T07:32:00-08:00                                            
#> <span style="color: #6F7783;">10 | </span>                                                                           
#> <span style="color: #6F7783;">ℹ 15 more lines</span>                                                                 
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}


Insert a key-value pair into a table:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(toml, "owner") |> ts_tree_insert(key = "dpt", "Engineering")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (24 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span># This is a TOML document                                                  
#> <span style="color: #6F7783;"> 2 | </span>                                                                           
#> <span style="color: #6F7783;"> 3 | </span>title = "TOML Example"                                                     
#> <span style="color: #6F7783;"> 4 | </span>                                                                           
#> <span style="color: #6F7783;"> 5 | </span>[owner]                                                                    
#> <span style="color: #6F7783;"> 6 | </span>name = "Tom Preston-Werner"                                                
#> <span style="color: #6F7783;"> 7 | </span>dob = 1979-05-27T07:32:00-08:00                                            
#> <span style="color: #6F7783;"> 8 | </span>dpt = "Engineering"                                                        
#> <span style="color: #6F7783;"> 9 | </span>                                                                           
#> <span style="color: #6F7783;">10 | </span>[database]                                                                 
#> <span style="color: #6F7783;">ℹ 14 more lines</span>                                                                 
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}


Insert an element into an array:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(toml, "database", "ports") |>
  ts_tree_insert(10000L) |>
  print(n = 100)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (23 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span># This is a TOML document                                                  
#> <span style="color: #6F7783;"> 2 | </span>                                                                           
#> <span style="color: #6F7783;"> 3 | </span>title = "TOML Example"                                                     
#> <span style="color: #6F7783;"> 4 | </span>                                                                           
#> <span style="color: #6F7783;"> 5 | </span>[owner]                                                                    
#> <span style="color: #6F7783;"> 6 | </span>name = "Tom Preston-Werner"                                                
#> <span style="color: #6F7783;"> 7 | </span>dob = 1979-05-27T07:32:00-08:00                                            
#> <span style="color: #6F7783;"> 8 | </span>                                                                           
#> <span style="color: #6F7783;"> 9 | </span>[database]                                                                 
#> <span style="color: #6F7783;">10 | </span>enabled = true                                                             
#> <span style="color: #6F7783;">11 | </span>ports = [ 8000, 8001, 8002, 10000 ]                                        
#> <span style="color: #6F7783;">12 | </span>data = [ ["delta", "phi"], [3.14] ]                                        
#> <span style="color: #6F7783;">13 | </span>temp_targets = \{ cpu = 79.5, case = 72.0 \}                                 
#> <span style="color: #6F7783;">14 | </span>                                                                           
#> <span style="color: #6F7783;">15 | </span>[servers]                                                                  
#> <span style="color: #6F7783;">16 | </span>                                                                           
#> <span style="color: #6F7783;">17 | </span>[servers.alpha]                                                            
#> <span style="color: #6F7783;">18 | </span>ip = "10.0.0.1"                                                            
#> <span style="color: #6F7783;">19 | </span>role = "frontend"                                                          
#> <span style="color: #6F7783;">20 | </span>                                                                           
#> <span style="color: #6F7783;">21 | </span>[servers.beta]                                                             
#> <span style="color: #6F7783;">22 | </span>ip = "10.0.0.2"                                                            
#> <span style="color: #6F7783;">23 | </span>role = "backend"                                                           
</pre></div>
}}


Insert a new table into the document:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_insert(toml, key = "new_table", list(a = 1, b = 2)) |>
  print(n = 100)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (27 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span># This is a TOML document                                                  
#> <span style="color: #6F7783;"> 2 | </span>                                                                           
#> <span style="color: #6F7783;"> 3 | </span>title = "TOML Example"                                                     
#> <span style="color: #6F7783;"> 4 | </span>                                                                           
#> <span style="color: #6F7783;"> 5 | </span>[owner]                                                                    
#> <span style="color: #6F7783;"> 6 | </span>name = "Tom Preston-Werner"                                                
#> <span style="color: #6F7783;"> 7 | </span>dob = 1979-05-27T07:32:00-08:00                                            
#> <span style="color: #6F7783;"> 8 | </span>                                                                           
#> <span style="color: #6F7783;"> 9 | </span>[database]                                                                 
#> <span style="color: #6F7783;">10 | </span>enabled = true                                                             
#> <span style="color: #6F7783;">11 | </span>ports = [ 8000, 8001, 8002 ]                                               
#> <span style="color: #6F7783;">12 | </span>data = [ ["delta", "phi"], [3.14] ]                                        
#> <span style="color: #6F7783;">13 | </span>temp_targets = \{ cpu = 79.5, case = 72.0 \}                                 
#> <span style="color: #6F7783;">14 | </span>                                                                           
#> <span style="color: #6F7783;">15 | </span>[servers]                                                                  
#> <span style="color: #6F7783;">16 | </span>                                                                           
#> <span style="color: #6F7783;">17 | </span>[servers.alpha]                                                            
#> <span style="color: #6F7783;">18 | </span>ip = "10.0.0.1"                                                            
#> <span style="color: #6F7783;">19 | </span>role = "frontend"                                                          
#> <span style="color: #6F7783;">20 | </span>                                                                           
#> <span style="color: #6F7783;">21 | </span>[servers.beta]                                                             
#> <span style="color: #6F7783;">22 | </span>ip = "10.0.0.2"                                                            
#> <span style="color: #6F7783;">23 | </span>role = "backend"                                                           
#> <span style="color: #6F7783;">24 | </span>                                                                           
#> <span style="color: #6F7783;">25 | </span>[new_table]                                                                
#> <span style="color: #6F7783;">26 | </span>a = 1.0                                                                    
#> <span style="color: #6F7783;">27 | </span>b = 2.0                                                                    
</pre></div>
}}


Insert a new array of tables into the document:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_insert(toml, key = "new_aot",
  list(list(x = 1), list(x = 2))) |>
  print(n = 100)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (29 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span># This is a TOML document                                                  
#> <span style="color: #6F7783;"> 2 | </span>                                                                           
#> <span style="color: #6F7783;"> 3 | </span>title = "TOML Example"                                                     
#> <span style="color: #6F7783;"> 4 | </span>                                                                           
#> <span style="color: #6F7783;"> 5 | </span>[owner]                                                                    
#> <span style="color: #6F7783;"> 6 | </span>name = "Tom Preston-Werner"                                                
#> <span style="color: #6F7783;"> 7 | </span>dob = 1979-05-27T07:32:00-08:00                                            
#> <span style="color: #6F7783;"> 8 | </span>                                                                           
#> <span style="color: #6F7783;"> 9 | </span>[database]                                                                 
#> <span style="color: #6F7783;">10 | </span>enabled = true                                                             
#> <span style="color: #6F7783;">11 | </span>ports = [ 8000, 8001, 8002 ]                                               
#> <span style="color: #6F7783;">12 | </span>data = [ ["delta", "phi"], [3.14] ]                                        
#> <span style="color: #6F7783;">13 | </span>temp_targets = \{ cpu = 79.5, case = 72.0 \}                                 
#> <span style="color: #6F7783;">14 | </span>                                                                           
#> <span style="color: #6F7783;">15 | </span>[servers]                                                                  
#> <span style="color: #6F7783;">16 | </span>                                                                           
#> <span style="color: #6F7783;">17 | </span>[servers.alpha]                                                            
#> <span style="color: #6F7783;">18 | </span>ip = "10.0.0.1"                                                            
#> <span style="color: #6F7783;">19 | </span>role = "frontend"                                                          
#> <span style="color: #6F7783;">20 | </span>                                                                           
#> <span style="color: #6F7783;">21 | </span>[servers.beta]                                                             
#> <span style="color: #6F7783;">22 | </span>ip = "10.0.0.2"                                                            
#> <span style="color: #6F7783;">23 | </span>role = "backend"                                                           
#> <span style="color: #6F7783;">24 | </span>                                                                           
#> <span style="color: #6F7783;">25 | </span>[[new_aot]]                                                                
#> <span style="color: #6F7783;">26 | </span>x = 1.0                                                                    
#> <span style="color: #6F7783;">27 | </span>                                                                           
#> <span style="color: #6F7783;">28 | </span>[[new_aot]]                                                                
#> <span style="color: #6F7783;">29 | </span>x = 2.0                                                                    
</pre></div>
}}

}

\subsection{Update elements}{

Update an existing element:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(toml, "title") |> ts_tree_update("A New TOML Example")
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (23 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span># This is a TOML document                                                  
#> <span style="color: #6F7783;"> 2 | </span>                                                                           
#> <span style="color: #6F7783;"> 3 | </span>title = "A New TOML Example"                                               
#> <span style="color: #6F7783;"> 4 | </span>                                                                           
#> <span style="color: #6F7783;"> 5 | </span>[owner]                                                                    
#> <span style="color: #6F7783;"> 6 | </span>name = "Tom Preston-Werner"                                                
#> <span style="color: #6F7783;"> 7 | </span>dob = 1979-05-27T07:32:00-08:00                                            
#> <span style="color: #6F7783;"> 8 | </span>                                                                           
#> <span style="color: #6F7783;"> 9 | </span>[database]                                                                 
#> <span style="color: #6F7783;">10 | </span>enabled = true                                                             
#> <span style="color: #6F7783;">ℹ 13 more lines</span>                                                                 
#> <span style="color: #6F7783;">ℹ Use `print(n = ...)` to see more lines</span>                                        
</pre></div>
}}


Update multiple elements at once:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ts_tree_select(toml, "servers", c("alpha", "beta"), "ip") |>
  ts_tree_update("192.168.1.23") |>
  print(n = 100)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #B9C0CB;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> <span style="color: #6F7783;"># toml (23 lines)</span>                                                               
#> <span style="color: #6F7783;"> 1 | </span># This is a TOML document                                                  
#> <span style="color: #6F7783;"> 2 | </span>                                                                           
#> <span style="color: #6F7783;"> 3 | </span>title = "TOML Example"                                                     
#> <span style="color: #6F7783;"> 4 | </span>                                                                           
#> <span style="color: #6F7783;"> 5 | </span>[owner]                                                                    
#> <span style="color: #6F7783;"> 6 | </span>name = "Tom Preston-Werner"                                                
#> <span style="color: #6F7783;"> 7 | </span>dob = 1979-05-27T07:32:00-08:00                                            
#> <span style="color: #6F7783;"> 8 | </span>                                                                           
#> <span style="color: #6F7783;"> 9 | </span>[database]                                                                 
#> <span style="color: #6F7783;">10 | </span>enabled = true                                                             
#> <span style="color: #6F7783;">11 | </span>ports = [ 8000, 8001, 8002 ]                                               
#> <span style="color: #6F7783;">12 | </span>data = [ ["delta", "phi"], [3.14] ]                                        
#> <span style="color: #6F7783;">13 | </span>temp_targets = \{ cpu = 79.5, case = 72.0 \}                                 
#> <span style="color: #6F7783;">14 | </span>                                                                           
#> <span style="color: #6F7783;">15 | </span>[servers]                                                                  
#> <span style="color: #6F7783;">16 | </span>                                                                           
#> <span style="color: #6F7783;">17 | </span>[servers.alpha]                                                            
#> <span style="color: #6F7783;">18 | </span>ip = "192.168.1.23"                                                        
#> <span style="color: #6F7783;">19 | </span>role = "frontend"                                                          
#> <span style="color: #6F7783;">20 | </span>                                                                           
#> <span style="color: #6F7783;">21 | </span>[servers.beta]                                                             
#> <span style="color: #6F7783;">22 | </span>ip = "192.168.1.23"                                                        
#> <span style="color: #6F7783;">23 | </span>role = "backend"                                                           
</pre></div>
}}

}
}
