% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-toml.R
\name{save_toml}
\alias{save_toml}
\title{Write a tstoml object to a file}
\usage{
save_toml(toml, file = NULL)
}
\arguments{
\item{toml}{tstoml object.}

\item{file}{File or connection to write to. Both binary and text
connections are supported. Use \code{stdout()} to output to the screen.}
}
\value{
Nothing.
}
\description{
Write a tstoml object to a file
}
\seealso{
\code{\link[=load_toml]{load_toml()}} to create a tstoml object from a TOML file or
string. \code{\link[=unserialize_selected]{unserialize_selected()}} obtain a TOML string for a tstoml
object.
}
