% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\alias{[[.tstoml}
\alias{select_refine}
\title{Select elements in a tstoml object}
\usage{
select(toml, ...)

\method{[[}{tstoml}(x, i, ...)

select_refine(toml, ...)
}
\arguments{
\item{x, toml}{tstoml object.}

\item{i, ...}{Selectors, see below.}
}
\description{
Select elements in a tstoml object
}
