% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize-toml.R
\name{serialize_toml}
\alias{serialize_toml}
\title{Serialize an R object to TOML}
\usage{
serialize_toml(obj, file = NULL, collapse = FALSE, options = NULL)
}
\arguments{
\item{obj}{R object to serialize.}

\item{file}{If not \code{NULL} then the result if written to this file.}

\item{collapse}{If \code{file} is \code{NULL} then whether to return a character
scalar or a character vector.}

\item{options}{A named list of \code{tstoml} options, see
\code{\link[=tstoml_options]{tstoml_options()}}.}
}
\value{
If \code{file} is \code{NULL} then a character scalar (\code{collapse} = TRUE)
or vector (\code{collapse} = FALSE). If \code{file} is not \code{NULL} then nothing.
}
\description{
Create TOML from an R object.
Note that this function is not a generic serializer that can represent
any R object in TOML. Also, you cannot expect that \code{\link[=unserialize_toml]{unserialize_toml()}}
will do the exact inverse of \code{\link[=serialize_toml]{serialize_toml()}}.
}
\details{
See the examples below on how to create all possible TOML elements with
\code{\link[=serialize_toml]{serialize_toml()}}.
}
\seealso{
\code{\link[=unserialize_toml]{unserialize_toml()}} for the opposite.
}
