% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize-toml.R
\name{serialize_toml_value}
\alias{serialize_toml_value}
\title{Serialize a TOML value}
\usage{
serialize_toml_value(obj, options = NULL)
}
\arguments{
\item{obj}{R object to serialize as a TOML value.}

\item{options}{A named list of \code{tstoml} options, see
\code{\link[=tstoml_options]{tstoml_options()}}.}
}
\description{
Serialize a TOML value
}
