% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sitter.R
\name{sexpr_toml}
\alias{sexpr_toml}
\title{Show the syntax tree structure of a TOML file or string}
\usage{
sexpr_toml(file = NULL, text = NULL, ranges = NULL)
}
\arguments{
\item{file}{Path of a TOML file. Use either \code{file} or \code{text}.}

\item{text}{TOML string. Use either \code{file} or \code{text}.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}
}
\description{
Show the syntax tree structure of a TOML file or string
}
\examples{
sexpr_toml(
  text = "[owner]\nname = \"Tom Preston-Werner\"\nemail = \"tom@github.com\""
)
}
