% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-sitter.R
\name{token_table}
\alias{token_table}
\title{Get the token table of a TOML file or string}
\usage{
token_table(
  file = NULL,
  text = NULL,
  ranges = NULL,
  fail_on_parse_error = TRUE,
  options = NULL
)
}
\arguments{
\item{file}{Path of a TOML file. Use either \code{file} or \code{text}.}

\item{text}{TOML string. Use either \code{file} or \code{text}.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}

\item{fail_on_parse_error}{Logical, whether to error if there are
parse errors in the toml document. Default is \code{TRUE}.}

\item{options}{List of options, see \code{\link[=tstoml_options]{tstoml_options()}}. This argument
must be named and cannot be abbreviated.}
}
\value{
A data frame with one row per token, and columns:
\itemize{
\item \code{id}: integer, the id of the token.
\item \code{parent}: integer, the id of the parent token. The root token has
parent \code{NA}
\item \code{field_name}: character, the field name of the token in its parent.
\item \code{type}: character, the type of the token.
\item \code{code}: character, the actual code of the token.
\item \code{start_byte}, \code{end_byte}: integer, the byte positions of the token
in the input.
\item \code{start_row}, \code{start_column}, \code{end_row}, \code{end_column}: integer, the
position of the token in the input.
\item \code{is_missing}: logical, whether the token is a missing token added by
the parser to recover from errors.
\item \code{has_error}: logical, whether the token has a parse error.
\item \code{children}: list of integer vectors, the ids of the children tokens.
}
}
\description{
Get the token table of a TOML file or string
}
\examples{
token_table(text = "{ \"a\": true, \"b\": [1, 2, 3] }")
}
