% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-format-toml.R
\name{ts_format_toml}
\alias{ts_format_toml}
\title{Format a TOML file}
\usage{
ts_format_toml(file = NULL, text = NULL, options = NULL)
}
\arguments{
\item{file}{Path to a TOML file. Use one of \code{file} or \code{text}.}

\item{text}{TOML content as a raw vector or a character vector. Use one
of \code{file} or \code{text}.}

\item{options}{Named list of parsing and formatting options, see
\link[=tstoml_options]{tstoml options}.}
}
\description{
Format a TOML file
}
