% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-delete.R
\name{ts_tree_delete.ts_tree_toml}
\alias{ts_tree_delete.ts_tree_toml}
\title{Delete selected elements from a tstoml object}
\usage{
\method{ts_tree_delete}{ts_tree_toml}(tree, ...)
}
\arguments{
\item{tree}{tstoml object.}

\item{...}{Reserved for future use.}
}
\value{
Modified tstoml object.
}
\description{
The formatting of the rest of TOML document is kept as is. Comments
appearing inside the deleted elements are also deleted. Other comments
are left as is.
}
\details{
If \code{toml} has no selection then the the whole document is deleted.
If \code{toml} has an empty selection, then nothing is delted.
}
\examples{
library(tsitter)
toml <- ts_parse_toml(text = toml_example_text())
toml

toml |> ts_tree_select("owner", "name") |> ts_tree_delete()
toml |> ts_tree_select("owner") |> ts_tree_delete()
}
\keyword{internal}
