% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-format-toml.R
\name{ts_tree_format.ts_tree_jsonc}
\alias{ts_tree_format.ts_tree_jsonc}
\title{Format the selected TOML elements}
\usage{
\method{ts_tree_format}{ts_tree_jsonc}(tree, options = NULL, ...)
}
\arguments{
\item{tree}{tstoml object.}

\item{options}{Named list of parsing options, see
\link[=tstoml_options]{tstoml options}.}

\item{...}{Reserved for future use.}
}
\value{
The updated tstoml object.
}
\description{
Format the selected TOML elements
}
\details{
If \code{tree} does not have a selection, then all of it is formatted.
If \code{tree} has an empty selection, then nothing happens.
}
