% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-query.R
\name{ts_tree_query.ts_tree_toml}
\alias{ts_tree_query.ts_tree_toml}
\title{Run tree-sitter queries on TOML tree-sitter trees}
\usage{
\method{ts_tree_query}{ts_tree_toml}(tree, query)
}
\arguments{
\item{tree}{\eval{tsitter:::doc_insert("ts_tree_query_param_tree", "tsjtoml")}}

\item{query}{\eval{tsitter:::doc_insert("ts_tree_query_param_query", "tsjtoml")}}
}
\value{
\eval{tsitter:::doc_insert("ts_tree_query_return", "tsjtoml")}
}
\description{
\eval{tsitter:::doc_insert("ts_tree_query_description", "tsjtoml")}
}
\details{
\eval{tsitter:::doc_insert("ts_tree_query_details", "tsjtoml")}
\eval{tsitter:::doc_extra()}
}
\examples{
# Select all numbers in a TOML document ------------------------------------
library(tsitter)
toml <- tstoml::ts_parse_toml(
  'a = 1\nb = [10.0, 20, 30]\nc = { c1 = true, c2 = 100 }'
)
toml |> ts_tree_query("[(float) (integer)] @number")
}
