% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select.ts_tree_toml}
\alias{ts_tree_select.ts_tree_toml}
\title{Select parts of a TOML tree-sitter tree}
\usage{
\method{ts_tree_select}{ts_tree_toml}(tree, ..., refine = FALSE)
}
\arguments{
\item{tree}{\eval{tsitter:::doc_insert("tsitter::ts_tree_select_param_tree", "tstoml")}}

\item{...}{\eval{tsitter:::doc_insert("tsitter::ts_tree_select_param_dots", "tstoml")}}

\item{refine}{\eval{tsitter:::doc_insert("tsitter::ts_tree_select_param_refine", "tstoml")}}
}
\value{
\eval{tsitter:::doc_insert("tsitter::ts_tree_select_return", "tstoml")}
}
\description{
\eval{tsitter:::doc_insert("tsitter::ts_tree_select_description", "tstoml")}

This is the S3 method of the \code{\link[tsitter:ts_tree_select]{tsitter::ts_tree_select()}} generic,
for \link{tstoml} objects.
}
\details{
\eval{tsitter:::doc_insert("tsitter::ts_tree_select_details", "tstoml")}
\eval{tsitter:::doc_insert("tstoml::ts_tree_select_examples", "tstoml")}
\eval{tsitter:::doc_extra()}
}
