% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-select.R
\name{ts_tree_select.tstoml}
\alias{ts_tree_select.tstoml}
\title{Select parts of a TOML tree-sitter tree}
\usage{
\method{ts_tree_select}{tstoml}(tree, ..., refine = FALSE)
}
\arguments{
\item{tree}{\eval{ts:::doc_insert("ts::ts_tree_select_param_tree", "tstoml")}}

\item{...}{\eval{ts:::doc_insert("ts::ts_tree_select_param_dots", "tstoml")}}

\item{refine}{\eval{ts:::doc_insert("ts::ts_tree_select_param_refine", "tstoml")}}
}
\value{
\eval{ts:::doc_insert("ts::ts_tree_select_return", "tstoml")}
}
\description{
\eval{ts:::doc_insert("ts::ts_tree_select_description", "tstoml")}

This is the S3 method of the \code{\link[ts:ts_tree_select]{ts::ts_tree_select()}} generic,
for \link{tstoml} objects.
}
\details{
\eval{ts:::doc_insert("ts::ts_tree_select_details", "tstoml")}
\eval{ts:::doc_insert("tstoml::ts_tree_select_examples", "tstoml")}
\eval{ts:::doc_extra()}
}
