% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unserialize-toml.R
\name{unserialize_selected}
\alias{unserialize_selected}
\title{Unserialize selected elements from a tstoml object}
\usage{
unserialize_selected(toml)
}
\arguments{
\item{toml}{tstoml object.}
}
\description{
Unserialize selected elements from a tstoml object
}
