% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-selected.R
\name{update_selected}
\alias{update_selected}
\title{Replace selected TOML values with new content}
\usage{
update_selected(toml, new, options = NULL)
}
\arguments{
\item{toml}{A tstoml object}

\item{new}{A R object to be converted to TOML using \code{\link[=serialize_toml_value]{serialize_toml_value()}}
and used as the new value.}

\item{options}{A named list of \code{tstoml} options, see
\code{\link[=tstoml_options]{tstoml_options()}}. Passed to \code{\link[=serialize_toml_value]{serialize_toml_value()}}.}
}
\description{
Replace all selected elements with a new element. If \code{toml} has no
selection then the whole document is replaced. If \code{toml} has an empty
selection, then nothing happens.
}
