% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-tree-delete.R
\name{delete_selected}
\alias{delete_selected}
\title{Delete selected elements from a tstoml object}
\usage{
delete_selected(toml)
}
\arguments{
\item{toml}{tstoml object.}
}
\value{
Modified tstoml object.
}
\description{
The formatting of the rest of TOML document is kept as is. Comments
appearing inside the deleted elements are also deleted. Other comments
are left as is.
}
\details{
If \code{toml} has no selection then the the whole document is deleted.
If \code{toml} has an empty selection, then nothing is delted.
}
\examples{
toml <- load_toml(text = toml_example_text())
toml

toml |> select("owner", "name") |> delete_selected()
toml |> select("owner") |> delete_selected()
}
