% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-format-toml.R
\name{format_selected}
\alias{format_selected}
\title{Format the selected TOML elements}
\usage{
format_selected(toml, options = NULL)
}
\arguments{
\item{toml}{tstoml object.}
}
\value{
The updated tstoml object.
}
\description{
Format the selected TOML elements
}
\details{
If \code{toml} does not have a selection, then all of it is formatted.
If \code{toml} has an empty selection, then nothing happens.
}
