% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_query}
\alias{select_query}
\title{Select the nodes matching a tree-sitter query in a tstoml object}
\usage{
select_query(toml, query, captures = NULL)
}
\arguments{
\item{toml}{tstoml object.}

\item{query}{String, a tree-sitter query.}

\item{captures}{Optional character vector, the names of the captures to
select. By default all captured nodes are selected.}
}
\description{
Select the nodes matching a tree-sitter query in a tstoml object
}
