% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-serialize-toml.R
\name{serialize_toml}
\alias{serialize_toml}
\alias{ts_toml_table}
\alias{ts_toml_inline_table}
\alias{ts_toml_array}
\alias{ts_toml_array_of_tables}
\title{Serialize an R object to TOML}
\usage{
serialize_toml(obj, file = NULL, collapse = FALSE, options = NULL)

ts_toml_table(...)

ts_toml_inline_table(...)

ts_toml_array(...)

ts_toml_array_of_tables(...)
}
\arguments{
\item{obj}{R object to serialize.}

\item{file}{If not \code{NULL} then the result if written to this file.}

\item{collapse}{If \code{file} is \code{NULL} then whether to return a character
scalar or a character vector.}

\item{options}{A named list of \code{tstoml} options, see
\code{\link[=tstoml_options]{tstoml_options()}}.}

\item{...}{Elements of the TOML table, array, inline table or array
of tables.}
}
\value{
If \code{file} is \code{NULL} then a character scalar (\code{collapse} = TRUE)
or vector (\code{collapse} = FALSE). If \code{file} is not \code{NULL} then nothing.
}
\description{
Create TOML from an R object.
Note that this function is not a generic serializer that can represent
any R object in TOML. Also, you cannot expect that \code{\link[=unserialize_toml]{unserialize_toml()}}
will do the exact inverse of \code{\link[=serialize_toml]{serialize_toml()}}.
}
\details{
See the examples below on how to create all possible TOML elements with
\code{\link[=serialize_toml]{serialize_toml()}}.

Use \code{ts_toml_table()} to make a list to be serialized as a TOML
table. All named lists are serialized as TOML tables by default, so
this is only readability.

Use \code{ts_toml_inline_table()} to make a list to be serialized
as an inline TOML table. By default named lists are serialized as
regular TOML tables, if possible.

Use \code{ts_toml_array()} to make a list to be serialized as a TOML
array. By default un-named lists are serialized as arrays, unless they
are lists of named lists (which are serialized as arrays of tables).
Use this funtion to mark any list to be serialized as an array.

Use \code{ts_toml_array_of_tables()} to make a list to be serialized
as a TOML array of tables. It must a list of named lists. By default
lists of named lists are serialized as arrays of tables, so this is
only for readability.
}
\seealso{
\code{\link[=unserialize_toml]{unserialize_toml()}} for the opposite.
}
