% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-parse-toml.R
\name{ts_parse_toml}
\alias{ts_parse_toml}
\alias{ts_read_toml}
\title{TODO}
\usage{
ts_parse_toml(text, ranges = NULL, fail_on_parse_error = TRUE, options = NULL)

ts_read_toml(file, ranges = NULL, fail_on_parse_error = TRUE, options = NULL)
}
\description{
TODO
}
