% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-parse-toml.R
\name{ts_parse_toml}
\alias{ts_parse_toml}
\alias{ts_read_toml}
\title{Parse a TOML file or string into a ts_tree_toml object}
\usage{
ts_parse_toml(text, ranges = NULL, fail_on_parse_error = TRUE, options = NULL)

ts_read_toml(file, ranges = NULL, fail_on_parse_error = TRUE, options = NULL)
}
\arguments{
\item{text}{String. Use either \code{file} or \code{text}, but not both.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}

\item{fail_on_parse_error}{Logical, whether to error if there are
parse errors in the document. Default is \code{TRUE}.}

\item{options}{Named list of parsing options, see
\link[=tstoml_options]{tstoml options}.}

\item{file}{Path of a file. Use either \code{file} or \code{text}, but not both.}
}
\value{
A \code{ts_tree_toml} object.
}
\description{
Parse a TOML file or string into a ts_tree_toml object
}
