% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-unserialize-toml.R
\name{ts_unserialize_toml}
\alias{ts_unserialize_toml}
\title{Unserialize TOML to R objects}
\usage{
ts_unserialize_toml(file = NULL, text = NULL, ranges = NULL, options = NULL)
}
\arguments{
\item{file}{\Sexpr[stage=render,results=rd]{ts:::doc_insert("ts::ts_tree_new_param_file", "ts")}}

\item{text}{\Sexpr[stage=render,results=rd]{ts:::doc_insert("ts::ts_tree_new_param_text", "ts")}}

\item{ranges}{\Sexpr[stage=render,results=rd]{ts:::doc_insert("ts::ts_tree_new_param_ranges", "ts")}}

\item{options}{Named list of parsing options, see
\link[=tstoml_options]{tstoml options}.}
}
\description{
Unserialize TOML to R objects
}
