% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstoml-package.R
\docType{package}
\name{tstoml-package}
\alias{tstoml}
\alias{tstoml-package}
\title{tstoml: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
