% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstoml-package.R
\docType{package}
\name{tstoml-package}
\alias{tstoml}
\alias{tstoml-package}
\title{tstoml: Edit TOML files}
\description{
Extract and manipulate parts of TOML files without touching the formatting and comments in other parts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaborcsardi/tstoml}
  \item \url{https://gaborcsardi.github.io/tstoml/}
  \item Report bugs at \url{https://github.com/gaborcsardi/tstoml/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
  \item  tree-sitter-toml authors [contributor]
}

}
\keyword{internal}
