% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-unserialize-toml.R
\name{unserialize_toml}
\alias{unserialize_toml}
\title{Unserialize TOML to R objects}
\usage{
unserialize_toml(file = NULL, text = NULL, ranges = NULL, options = NULL)
}
\arguments{
\item{file}{Path of a file. Use either \code{file} or \code{text}, but not both.}

\item{text}{String. Use either \code{file} or \code{text}, but not both.}

\item{ranges}{Can be used to parse part(s) of the input. It must be a
data frame with integer columns \code{start_row}, \code{start_col}, \code{end_row},
\code{end_col}, \code{start_byte}, \code{end_byte}, in this order.}

\item{options}{Named list of parsing options, see
\link[=tstoml_options]{tstoml options}.}
}
\description{
Unserialize TOML to R objects
}
