% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-unserialize-toml.R
\name{unserialize_toml}
\alias{unserialize_toml}
\title{Unserialize TOML to R objects}
\usage{
unserialize_toml(file = NULL, text = NULL, ranges = NULL, options = NULL)
}
\description{
Unserialize TOML to R objects
}
